fix(map): fix type mismatch between selected circle ID and last fitted circle ID

+++ android-native/app/src/main/java/com/estoyok/app/features/tracking/presentation/MapaScreen.kt
@@ -95,7 +95,7 @@ fun MapaScreen(
     }
 
     val scope = rememberCoroutineScope()
-    var lastFittedCircleId by remember { mutableStateOf<Long?>(null) }
+    var lastFittedCircleId by remember { mutableStateOf<Int?>(null) }
 
     val fitAllMembers: () -> Unit = {
         val membersWithLocation = viewModel.selectedCircleMembers.filter { it.currentLocation != null }
